Remove hashes in favor of unique IDs
All hash functionality has been removed in favor of a simple base64
encode. The repository scheme `hashes` has been renamed to `ids`
and questions have been removed.

This allows the ids to be reverse-engineered back into questions
without the need to store questions. The removal of hashes was
predicated on the potential for collisions within the 32-bit int space
allotted, as well as not being able to decode a hash.