Rename DictionaryManager{} to ParityManger{}
The word Dictionary, is already associated with defining
words or being a lookup table of words. I should not redefine
it to mean "table of similarly grouped words".

The word Parity works better because it expresses a very specific
relationship. Having word-parity can mean to have a logical
grouping of words.