Add bold comment about `inquiryScheme` field order
Unfortunately, the field order matters when loading data
from a file into a scheme. This means if at any point we
decide to change the order of any properties, there needs to
be a legacy method added to update the existing database.

It goes without saying that if any fields are added, the same legacy
methods will need to be employed.